backend/requirements.txt
Django==4.2.9
djangorestframework==3.14.0
django-cors-headers==4.3.1
mysqlclient==2.2.4
groq==0.11.0
httpx==0.27.2
python-dotenv==1.0.1
Pillow==10.2.0
pydub==0.25.1
